+++ tests/classification/test_f_beta.py
@@ -32,8 +32,9 @@ from tests.classification.inputs import _input_multilabel_logits as _input_mlb_l
 from tests.classification.inputs import _input_multilabel_prob as _input_mlb_prob
 from tests.helpers import seed_all
 from tests.helpers.testers import NUM_BATCHES, NUM_CLASSES, THRESHOLD, MetricTester
-from torchmetrics import F1, FBeta, Metric
-from torchmetrics.functional import f1, fbeta
+from torchmetrics import F1Score, FBeta, Metric
+from torchmetrics.functional import f1_score as f1_score_pl
+from torchmetrics.functional import fbeta
 from torchmetrics.utilities.checks import _input_format_classification
 from torchmetrics.utilities.enums import AverageMethod
 
@@ -93,7 +94,7 @@ def _sk_fbeta_f1_multidim_multiclass(
     "metric_class, metric_fn",
     [
         (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
-        (F1, f1),
+        (F1Score, f1_score_pl),
     ],
 )
 @pytest.mark.parametrize(
@@ -129,7 +130,7 @@ def test_wrong_params(metric_class, metric_fn, average, mdmc_average, num_classe
     "metric_class, metric_fn",
     [
         (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
-        (F1, f1),
+        (F1Score, f1_score_pl),
     ],
 )
 def test_zero_division(metric_class, metric_fn):
@@ -151,7 +152,7 @@ def test_zero_division(metric_class, metric_fn):
     "metric_class, metric_fn",
     [
         (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
-        (F1, f1),
+        (F1Score, f1_score_pl),
     ],
 )
 def test_no_support(metric_class, metric_fn):
@@ -178,7 +179,13 @@ def test_no_support(metric_class, metric_fn):
     assert result_cl == result_fn == 0
 
 
-@pytest.mark.parametrize("metric_class, metric_fn", [(partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)), (F1, f1)])
+@pytest.mark.parametrize(
+    "metric_class, metric_fn",
+    [
+        (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
+        (F1Score, f1_score_pl),
+    ],
+)
 @pytest.mark.parametrize(
     "ignore_index, expected", [(None, torch.tensor([1.0, np.nan])), (0, torch.tensor([np.nan, np.nan]))]
 )
@@ -202,7 +209,10 @@ def test_class_not_present(metric_class, metric_fn, ignore_index, expected):
 
 @pytest.mark.parametrize(
     "metric_class, metric_fn, sk_fn",
-    [(partial(FBeta, beta=2.0), partial(fbeta, beta=2.0), partial(fbeta_score, beta=2.0)), (F1, f1, f1_score)],
+    [
+        (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0), partial(fbeta_score, beta=2.0)),
+        (F1Score, f1_score_pl, f1_score),
+    ],
 )
 @pytest.mark.parametrize("average", ["micro", "macro", None, "weighted", "samples"])
 @pytest.mark.parametrize("ignore_index", [None, 0])
@@ -388,7 +398,7 @@ _ml_k_preds = torch.tensor([[0.9, 0.2, 0.75], [0.1, 0.7, 0.8], [0.6, 0.1, 0.7]])
     "metric_class, metric_fn",
     [
         (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0)),
-        (F1, fbeta),
+        (F1Score, fbeta),
     ],
 )
 @pytest.mark.parametrize(
@@ -430,7 +440,10 @@ def test_top_k(
 @pytest.mark.parametrize("average", ["micro", "macro", "weighted"])
 @pytest.mark.parametrize(
     "metric_class, metric_functional, sk_fn",
-    [(partial(FBeta, beta=2.0), partial(fbeta, beta=2.0), partial(fbeta_score, beta=2.0)), (F1, f1, f1_score)],
+    [
+        (partial(FBeta, beta=2.0), partial(fbeta, beta=2.0), partial(fbeta_score, beta=2.0)),
+        (F1Score, f1_score_pl, f1_score),
+    ],
 )
 def test_same_input(metric_class, metric_functional, sk_fn, average, ignore_index):
     preds = _input_miss_class.preds