+++ tests/text/test_wer.py
@@ -11,8 +11,8 @@ if _JIWER_AVAILABLE:
 else:
     compute_measures = Callable
 
-from torchmetrics.functional.text.wer import wer
-from torchmetrics.text.wer import WER
+from torchmetrics.functional.text.wer import word_error_rate
+from torchmetrics.text.wer import WordErrorRate
 
 
 def _compute_wer_metric_jiwer(prediction: Union[str, List[str]], reference: Union[str, List[str]]):
@@ -36,7 +36,7 @@ class TestWER(TextTester):
             ddp=ddp,
             preds=preds,
             targets=targets,
-            metric_class=WER,
+            metric_class=WordErrorRate,
             sk_metric=_compute_wer_metric_jiwer,
             dist_sync_on_step=dist_sync_on_step,
         )
@@ -46,7 +46,7 @@ class TestWER(TextTester):
         self.run_functional_metric_test(
             preds,
             targets,
-            metric_functional=wer,
+            metric_functional=word_error_rate,
             sk_metric=_compute_wer_metric_jiwer,
         )
 
@@ -55,6 +55,6 @@ class TestWER(TextTester):
         self.run_differentiability_test(
             preds=preds,
             targets=targets,
-            metric_module=WER,
-            metric_functional=wer,
+            metric_module=WordErrorRate,
+            metric_functional=word_error_rate,
         )