+++ torchmetrics/audio/snr.py
@@ -12,14 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 from typing import Any, Callable, Optional
+from warnings import warn
 
 from torch import Tensor, tensor
 
-from torchmetrics.functional.audio.snr import snr
+from torchmetrics.functional.audio.snr import scale_invariant_signal_noise_ratio, snr
 from torchmetrics.metric import Metric
 
 
-class SNR(Metric):
+class SignalNoiseRatio(Metric):
     r"""Signal-to-noise ratio (SNR_):
 
     .. math::
@@ -57,12 +58,11 @@ class SNR(Metric):
 
     Example:
         >>> import torch
-        >>> from torchmetrics import SNR
+        >>> from torchmetrics import SignalNoiseRatio
         >>> target = torch.tensor([3.0, -0.5, 2.0, 7.0])
         >>> preds = torch.tensor([2.5, 0.0, 2.0, 8.0])
-        >>> snr = SNR()
-        >>> snr_val = snr(preds, target)
-        >>> snr_val
+        >>> snr = SignalNoiseRatio()
+        >>> snr(preds, target)
         tensor(16.1805)
 
     References:
@@ -109,3 +109,110 @@ class SNR(Metric):
     def compute(self) -> Tensor:
         """Computes average SNR."""
         return self.sum_snr / self.total
+
+
+class SNR(SignalNoiseRatio):
+    r"""Signal-to-noise ratio (SNR_):
+
+    .. deprecated:: v0.7
+        Use :class:`torchmetrics.SignalNoiseRatio`. Will be removed in v0.8.
+
+    Example:
+        >>> import torch
+        >>> snr = SNR()
+        >>> snr(torch.tensor([2.5, 0.0, 2.0, 8.0]), torch.tensor([3.0, -0.5, 2.0, 7.0]))
+        tensor(16.1805)
+
+    """
+
+    def __init__(
+        self,
+        zero_mean: bool = False,
+        compute_on_step: bool = True,
+        dist_sync_on_step: bool = False,
+        process_group: Optional[Any] = None,
+        dist_sync_fn: Optional[Callable[[Tensor], Tensor]] = None,
+    ) -> None:
+        warn("`SNR` was renamed to `SignalNoiseRatio` in v0.7 and it will be removed in v0.8", DeprecationWarning)
+        super().__init__(zero_mean, compute_on_step, dist_sync_on_step, process_group, dist_sync_fn)
+
+
+class ScaleInvariantSignalNoiseRatio(Metric):
+    """Scale-invariant signal-to-noise ratio (SI-SNR).
+
+    Forward accepts
+
+    - ``preds``: ``shape [...,time]``
+    - ``target``: ``shape [...,time]``
+
+    Args:
+        compute_on_step:
+            Forward only calls ``update()`` and returns None if this is set to False.
+        dist_sync_on_step:
+            Synchronize metric state across processes at each ``forward()``
+            before returning the value at the step.
+        process_group:
+            Specify the process group on which synchronization is called.
+        dist_sync_fn:
+            Callback that performs the allgather operation on the metric state. When `None`, DDP
+            will be used to perform the allgather.
+
+    Raises:
+        TypeError:
+            if target and preds have a different shape
+
+    Returns:
+        average si-snr value
+
+    Example:
+        >>> import torch
+        >>> from torchmetrics import ScaleInvariantSignalNoiseRatio
+        >>> target = torch.tensor([3.0, -0.5, 2.0, 7.0])
+        >>> preds = torch.tensor([2.5, 0.0, 2.0, 8.0])
+        >>> si_snr = ScaleInvariantSignalNoiseRatio()
+        >>> si_snr(preds, target)
+        tensor(15.0918)
+
+    References:
+        [1] Y. Luo and N. Mesgarani, "TaSNet: Time-Domain Audio Separation Network for Real-Time, Single-Channel Speech
+        Separation," 2018 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP), 2018, pp.
+        696-700, doi: 10.1109/ICASSP.2018.8462116.
+    """
+
+    is_differentiable = True
+    sum_si_snr: Tensor
+    total: Tensor
+    higher_is_better = True
+
+    def __init__(
+        self,
+        compute_on_step: bool = True,
+        dist_sync_on_step: bool = False,
+        process_group: Optional[Any] = None,
+        dist_sync_fn: Optional[Callable[[Tensor], Tensor]] = None,
+    ) -> None:
+        super().__init__(
+            compute_on_step=compute_on_step,
+            dist_sync_on_step=dist_sync_on_step,
+            process_group=process_group,
+            dist_sync_fn=dist_sync_fn,
+        )
+
+        self.add_state("sum_si_snr", default=tensor(0.0), dist_reduce_fx="sum")
+        self.add_state("total", default=tensor(0), dist_reduce_fx="sum")
+
+    def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
+        """Update state with predictions and targets.
+
+        Args:
+            preds: Predictions from model
+            target: Ground truth values
+        """
+        si_snr_batch = scale_invariant_signal_noise_ratio(preds=preds, target=target)
+
+        self.sum_si_snr += si_snr_batch.sum()
+        self.total += si_snr_batch.numel()
+
+    def compute(self) -> Tensor:
+        """Computes average SI-SNR."""
+        return self.sum_si_snr / self.total